misc fixes to make the ant task work better

+++ src/test/java/sagex/plugin/resolver/ant/SageDependenciesTest.java
@@ -14,8 +14,14 @@ public class SageDependenciesTest {
     @Test
     public void testExecute() throws Exception {
         File cacheDir=new File("target/cache/libs");
-        for (File f: cacheDir.listFiles()) {
-            f.delete();
+        if (cacheDir!=null) {
+            try {
+                for (File f : cacheDir.listFiles()) {
+                    f.delete();
+                }
+            } catch (Throwable t) {
+                // already deleted
+            }
         }
         File lastFM = new File(cacheDir, "last.fm-bindings.jar");
         SageDependencies sageDependencies = new SageDependencies();
@@ -24,7 +30,7 @@ public class SageDependenciesTest {
         sageDependencies.setDevPluginsXml(SageDependencies.class.getResource("SageTVPluginsDev.xml").toExternalForm());
         sageDependencies.setExtraJars("http://central.maven.org/maven2/com/squareup/retrofit/retrofit/1.9.0/retrofit-1.9.0.jar, http://central.maven.org/maven2/com/squareup/okhttp/okhttp/2.2.0/okhttp-2.2.0.jar");
         sageDependencies.execute();
-        for (String jar: new String[] {"last.fm-bindings.jar", "retrofit-1.9.0.jar", "okhttp-2.2.0.jar"}) {
+        for (String jar: new String[] {"last.fm-bindings.jar", "retrofit-1.9.0.jar", "okhttp-2.2.0.jar", "commons-io-2.4.jar"}) {
             File f = new File(cacheDir, jar);
             assertTrue("FILE Should Exist, but does not: " + f,f.exists());
         }
